実装終了

+++ db/schema.rb
@@ -75,8 +75,8 @@ ActiveRecord::Schema[7.0].define(version: 2023_12_21_091200) do
     t.string "encrypted_password", default: "", null: false
     t.date "birthday", null: false
     t.integer "sex_id", null: false
-    t.integer "height", null: false
-    t.integer "weight", null: false
+    t.float "height", null: false
+    t.float "weight", null: false
     t.string "reset_password_token"
     t.datetime "reset_password_sent_at"
     t.datetime "remember_created_at"
